feat: fix native provider for transcription
